fix: Rename column to selectedcolumn in DB (#90)
`column` seems to be a reserved keyword in postgresql. really cool that
moodle never warned me of this.

anyway, this should fix EDUPL-19 if I'm correct.

this PR also:
- reverts previous change regarding extra debugging on this DB request
(it did its job and will be superseded by #89 later)
- reverts workarounds I put in place precisely so that I wouldn't have
to rename the field (postgres seems to have "interesting" behaviour
around inserting stuff without IDs)

do note that inside the code, the field in the \*model\* is still called
"column" - this is on one hand for brevity, and on the other so I don't
have to refactor as much.